@@ -1,36 +1,31 @@
 cmake_minimum_required(VERSION 2.8.12)
 project(iiwa7_ikfast_iiwa_arm_plugin)
 
-add_compile_options(-std=c++11)
+add_compile_options(-std=c++14)
 
 find_package(catkin REQUIRED COMPONENTS
   moveit_core
   pluginlib
   roscpp
-  tf_conversions
+  tf2_kdl
+  tf2_eigen
+  eigen_conversions
 )
-
-include_directories(${catkin_INCLUDE_DIRS})
-
-catkin_package(
-  LIBRARIES
-  CATKIN_DEPENDS
-    moveit_core
-    pluginlib
-    roscpp
-    tf_conversions
-)
-
-include_directories(include)
-
-set(IKFAST_LIBRARY_NAME iiwa7_iiwa_arm_moveit_ikfast_plugin)
-
 find_package(LAPACK REQUIRED)
 
+include_directories(${catkin_INCLUDE_DIRS} include)
+
+catkin_package()
+
+set(IKFAST_LIBRARY_NAME iiwa7_iiwa_arm_moveit_ikfast_plugin)
 add_library(${IKFAST_LIBRARY_NAME} src/iiwa7_iiwa_arm_ikfast_moveit_plugin.cpp)
 target_link_libraries(${IKFAST_LIBRARY_NAME} ${catkin_LIBRARIES} ${Boost_LIBRARIES} ${LAPACK_LIBRARIES})
+# suppress warnings about unused variables in OpenRave's solver code
+target_compile_options(${IKFAST_LIBRARY_NAME} PRIVATE -Wno-unused-variable)
 
-install(TARGETS ${IKFAST_LIBRARY_NAME} LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION})
+install(TARGETS
+  ${IKFAST_LIBRARY_NAME}
+  LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION})
 
 install(
   FILES
