@@ -34,22 +34,22 @@
  *********************************************************************/
 
 /*
- * IKFast Plugin Template for moveit
+ * IKFast Kinematics Solver Plugin for MoveIt! wrapping an ikfast solver from OpenRAVE.
  *
- * AUTO-GENERATED by create_ikfast_moveit_plugin.py in arm_kinematics_tools
- * You should run create_ikfast_moveit_plugin.py to generate your own plugin.
+ * AUTO-GENERATED by create_ikfast_moveit_plugin.py in moveit_kinematics package.
  *
- * Creates a kinematics plugin using the output of IKFast from OpenRAVE.
- * This plugin and the move_group node can be used as a general
- * kinematics service, from within the moveit planning environment, or in
- * your own ROS node.
- *
+ * This file, including the ikfast cpp from OpenRAVE below, forms a MoveIt! kinematics plugin.
  */
 
 #include <ros/ros.h>
 #include <moveit/kinematics_base/kinematics_base.h>
-#include <urdf/model.h>
-#include <tf_conversions/tf_kdl.h>
+#include <moveit/robot_state/robot_state.h>
+#include <Eigen/Geometry>
+#include <tf2_kdl/tf2_kdl.h>
+#include <tf2_eigen/tf2_eigen.h>
+#include <eigen_conversions/eigen_kdl.h>
+
+using namespace moveit::core;
 
 // Need a floating point tolerance when checking joint limits, in case the joint starts at limit
 const double LIMIT_TOLERANCE = .0000001;
@@ -60,7 +60,7 @@ enum SEARCH_MODE
   OPTIMIZE_MAX_JOINT = 2
 };
 
-namespace ikfast_kinematics_plugin
+namespace iiwa7_iiwa_arm
 {
 #define IKFAST_NO_MAIN  // Don't include main() from IKFast
 
@@ -160,14 +160,32 @@ class IKFastKinematicsPlugin : public kinematics::KinematicsBase
   std::vector<std::string> link_names_;
   const size_t num_joints_;
   std::vector<int> free_params_;
-  bool active_;  // Internal variable that indicates whether solvers are configured and ready
+
+  // The ikfast and base frame are the start and end of the kinematic chain for which the
+  // IKFast analytic solution was generated.
+  const std::string IKFAST_TIP_FRAME_ = "iiwa_link_ee";
+  const std::string IKFAST_BASE_FRAME_ = "iiwa_link_0";
+
+  // The transform tip and base bool are set to true if this solver is used with a kinematic
+  // chain that extends beyond the ikfast tip and base frame. The solution will be valid so
+  // long as there are no active, passive, or mimic joints between either the ikfast_tip_frame
+  // and the tip_frame of the group or the ikfast_base_frame and the base_frame for the group.
+  bool tip_transform_required_;
+  bool base_transform_required_;
+
+  // We store the transform from the ikfast_base_frame to the group base_frame as well as the
+  // ikfast_tip_frame to the group tip_frame to transform input poses into the solver frame.
+  Eigen::Isometry3d chain_base_to_group_base_;
+  Eigen::Isometry3d group_tip_to_chain_tip_;
+
+  bool initialized_;  // Internal variable that indicates whether solvers are configured and ready
   const std::string name_{ "ikfast" };
 
-  const std::vector<std::string>& getJointNames() const
+  const std::vector<std::string>& getJointNames() const override
   {
     return joint_names_;
   }
-  const std::vector<std::string>& getLinkNames() const
+  const std::vector<std::string>& getLinkNames() const override
   {
     return link_names_;
   }
@@ -176,9 +194,9 @@ public:
   /** @class
    *  @brief Interface for an IKFast kinematics plugin
    */
-  IKFastKinematicsPlugin() : num_joints_(GetNumJoints()), active_(false)
+  IKFastKinematicsPlugin() : num_joints_(GetNumJoints()), initialized_(false)
   {
-    srand(time(NULL));
+    srand(time(nullptr));
     supported_methods_.push_back(kinematics::DiscretizationMethods::NO_DISCRETIZATION);
     supported_methods_.push_back(kinematics::DiscretizationMethods::ALL_DISCRETIZED);
     supported_methods_.push_back(kinematics::DiscretizationMethods::ALL_RANDOM_SAMPLED);
@@ -194,9 +212,10 @@ public:
    */
 
   // Returns the IK solution that is within joint limits closest to ik_seed_state
-  bool getPositionIK(const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state,
-                     std::vector<double>& solution, moveit_msgs::MoveItErrorCodes& error_code,
-                     const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const;
+  bool getPositionIK(
+      const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state, std::vector<double>& solution,
+      moveit_msgs::MoveItErrorCodes& error_code,
+      const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const override;
 
   /**
    * @brief Given a desired pose of the end-effector, compute the set joint angles solutions that are able to reach it.
@@ -215,7 +234,7 @@ public:
    */
   bool getPositionIK(const std::vector<geometry_msgs::Pose>& ik_poses, const std::vector<double>& ik_seed_state,
                      std::vector<std::vector<double>>& solutions, kinematics::KinematicsResult& result,
-                     const kinematics::KinematicsQueryOptions& options) const;
+                     const kinematics::KinematicsQueryOptions& options) const override;
 
   /**
    * @brief Given a desired pose of the end-effector, search for the joint angles required to reach it.
@@ -225,9 +244,10 @@ public:
    * @param ik_seed_state an initial guess solution for the inverse kinematics
    * @return True if a valid solution was found, false otherwise
    */
-  bool searchPositionIK(const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state, double timeout,
-                        std::vector<double>& solution, moveit_msgs::MoveItErrorCodes& error_code,
-                        const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const;
+  bool searchPositionIK(
+      const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state, double timeout,
+      std::vector<double>& solution, moveit_msgs::MoveItErrorCodes& error_code,
+      const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const override;
 
   /**
    * @brief Given a desired pose of the end-effector, search for the joint angles required to reach it.
@@ -238,10 +258,11 @@ public:
    * @param the distance that the redundancy can be from the current position
    * @return True if a valid solution was found, false otherwise
    */
-  bool searchPositionIK(const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state, double timeout,
-                        const std::vector<double>& consistency_limits, std::vector<double>& solution,
-                        moveit_msgs::MoveItErrorCodes& error_code,
-                        const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const;
+  bool searchPositionIK(
+      const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state, double timeout,
+      const std::vector<double>& consistency_limits, std::vector<double>& solution,
+      moveit_msgs::MoveItErrorCodes& error_code,
+      const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const override;
 
   /**
    * @brief Given a desired pose of the end-effector, search for the joint angles required to reach it.
@@ -251,10 +272,10 @@ public:
    * @param ik_seed_state an initial guess solution for the inverse kinematics
    * @return True if a valid solution was found, false otherwise
    */
-  bool searchPositionIK(const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state, double timeout,
-                        std::vector<double>& solution, const IKCallbackFn& solution_callback,
-                        moveit_msgs::MoveItErrorCodes& error_code,
-                        const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const;
+  bool searchPositionIK(
+      const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state, double timeout,
+      std::vector<double>& solution, const IKCallbackFn& solution_callback, moveit_msgs::MoveItErrorCodes& error_code,
+      const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const override;
 
   /**
    * @brief Given a desired pose of the end-effector, search for the joint angles required to reach it.
@@ -266,10 +287,11 @@ public:
    * @param consistency_limit the distance that the redundancy can be from the current position
    * @return True if a valid solution was found, false otherwise
    */
-  bool searchPositionIK(const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state, double timeout,
-                        const std::vector<double>& consistency_limits, std::vector<double>& solution,
-                        const IKCallbackFn& solution_callback, moveit_msgs::MoveItErrorCodes& error_code,
-                        const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const;
+  bool searchPositionIK(
+      const geometry_msgs::Pose& ik_pose, const std::vector<double>& ik_seed_state, double timeout,
+      const std::vector<double>& consistency_limits, std::vector<double>& solution,
+      const IKCallbackFn& solution_callback, moveit_msgs::MoveItErrorCodes& error_code,
+      const kinematics::KinematicsQueryOptions& options = kinematics::KinematicsQueryOptions()) const override;
 
   /**
    * @brief Given a set of joint angles and a set of links, compute their pose
@@ -280,7 +302,7 @@ public:
    * @return True if a valid solution was found, false otherwise
    */
   bool getPositionFK(const std::vector<std::string>& link_names, const std::vector<double>& joint_angles,
-                     std::vector<geometry_msgs::Pose>& poses) const;
+                     std::vector<geometry_msgs::Pose>& poses) const override;
 
   /**
    * @brief Sets the discretization value for the redundant joint.
@@ -291,22 +313,23 @@ public:
    *
    * @param discretization a map of joint indices and discretization value pairs.
    */
-  void setSearchDiscretization(const std::map<int, double>& discretization);
+  void setSearchDiscretization(const std::map<unsigned int, double>& discretization);
 
   /**
    * @brief Overrides the default method to prevent changing the redundant joints
    */
-  bool setRedundantJoints(const std::vector<unsigned int>& redundant_joint_indices);
+  bool setRedundantJoints(const std::vector<unsigned int>& redundant_joint_indices) override;
 
 private:
-  bool initialize(const std::string& robot_description, const std::string& group_name, const std::string& base_name,
-                  const std::string& tip_name, double search_discretization);
+  bool initialize(const moveit::core::RobotModel& robot_model, const std::string& group_name,
+                  const std::string& base_frame, const std::vector<std::string>& tip_frames,
+                  double search_discretization) override;
 
   /**
    * @brief Calls the IK solver from IKFast
    * @return The number of solutions found
    */
-  int solve(KDL::Frame& pose_frame, const std::vector<double>& vfree, IkSolutionList<IkReal>& solutions) const;
+  size_t solve(KDL::Frame& pose_frame, const std::vector<double>& vfree, IkSolutionList<IkReal>& solutions) const;
 
   /**
    * @brief Gets a specific solution from the set
@@ -319,10 +342,11 @@ private:
   void getSolution(const IkSolutionList<IkReal>& solutions, const std::vector<double>& ik_seed_state, int i,
                    std::vector<double>& solution) const;
 
-  double harmonize(const std::vector<double>& ik_seed_state, std::vector<double>& solution) const;
-  // void getOrderedSolutions(const std::vector<double> &ik_seed_state, std::vector<std::vector<double> >& solslist);
-  void getClosestSolution(const IkSolutionList<IkReal>& solutions, const std::vector<double>& ik_seed_state,
-                          std::vector<double>& solution) const;
+  /**
+   * @brief If the value is outside of min/max then it tries to +/- 2 * pi to put the value into the range
+   */
+  double enforceLimits(double val, double min, double max) const;
+
   void fillFreeParams(int count, int* array);
   bool getCount(int& count, const int& max_count, const int& min_count) const;
 
@@ -334,112 +358,114 @@ private:
   */
   bool sampleRedundantJoint(kinematics::DiscretizationMethod method, std::vector<double>& sampled_joint_vals) const;
 
+  /// Validate that we can compute a fixed transform between from and to links.
+  bool computeRelativeTransform(const std::string& from, const std::string& to, Eigen::Isometry3d& transform,
+                                bool& differs_from_identity);
+  /**
+  * @brief  Transforms the input pose to the correct frame for the solver. This assumes that the group includes the
+  * entire solver chain and that any joints outside of the solver chain within the group are are fixed.
+  * @param  ik_pose             The pose to be transformed which should be in the correct frame for the group.
+  * @param  ik_pose_chain       The ik_pose to be populated with the apropriate pose for the solver
+  */
+  void transformToChainFrame(const geometry_msgs::Pose& ik_pose, KDL::Frame& ik_pose_chain) const;
 };  // end class
 
-bool IKFastKinematicsPlugin::initialize(const std::string& robot_description, const std::string& group_name,
-                                        const std::string& base_name, const std::string& tip_name,
+bool IKFastKinematicsPlugin::computeRelativeTransform(const std::string& from, const std::string& to,
+                                                      Eigen::Isometry3d& transform, bool& differs_from_identity)
+{
+  RobotStatePtr robot_state;
+  robot_state.reset(new RobotState(robot_model_));
+  robot_state->setToDefaultValues();
+
+  auto* from_link = robot_state->getLinkModel(from);
+  auto* to_link = robot_state->getLinkModel(to);
+  if (!from_link)
+    ROS_ERROR_STREAM_NAMED(name_, "Could not find frame " << from);
+  if (!to_link)
+    ROS_ERROR_STREAM_NAMED(name_, "Could not find frame " << to);
+  if (!from_link || !to_link)
+    return false;
+
+  if (robot_state->getRobotModel()->getRigidlyConnectedParentLinkModel(from_link) !=
+      robot_state->getRobotModel()->getRigidlyConnectedParentLinkModel(to_link))
+  {
+    ROS_ERROR_STREAM_NAMED(name_, "Link frames " << from << " and " << to << " are not rigidly connected.");
+    return false;
+  }
+
+  transform = robot_state->getGlobalLinkTransform(from_link).inverse() * robot_state->getGlobalLinkTransform(to_link);
+  differs_from_identity = !transform.matrix().isIdentity();
+  return true;
+}
+
+bool IKFastKinematicsPlugin::initialize(const moveit::core::RobotModel& robot_model, const std::string& group_name,
+                                        const std::string& base_frame, const std::vector<std::string>& tip_frames,
                                         double search_discretization)
 {
-  setValues(robot_description, group_name, base_name, tip_name, search_discretization);
+  if (tip_frames.size() != 1)
+  {
+    ROS_ERROR_NAMED(name_, "Expecting exactly one tip frame.");
+    return false;
+  }
 
-  ros::NodeHandle node_handle("~/" + group_name);
+  storeValues(robot_model, group_name, base_frame, tip_frames, search_discretization);
 
-  std::string robot;
-  lookupParam("robot", robot, std::string());
+  // This IKFast solution was generated with IKFAST_TIP_FRAME_ and IKFAST_BASE_FRAME_.
+  // It is often the case that fixed joints are added to these links to model things like
+  // a robot mounted on a table or a robot with an end effector attached to the last link.
+  // To support these use cases, we store the transform from the IKFAST_BASE_FRAME_ to the
+  // base_frame_ and IKFAST_TIP_FRAME_ the tip_frame_ and transform to the input pose accordingly
+  if (!computeRelativeTransform(tip_frames_[0], IKFAST_TIP_FRAME_, group_tip_to_chain_tip_, tip_transform_required_) ||
+      !computeRelativeTransform(IKFAST_BASE_FRAME_, base_frame_, chain_base_to_group_base_, base_transform_required_))
+    return false;
 
   // IKFast56/61
   fillFreeParams(GetNumFreeParameters(), GetFreeParameters());
 
   if (free_params_.size() > 1)
   {
-    ROS_FATAL("Only one free joint parameter supported!");
+    ROS_ERROR_NAMED(name_, "Only one free joint parameter supported!");
     return false;
   }
   else if (free_params_.size() == 1)
   {
     redundant_joint_indices_.clear();
     redundant_joint_indices_.push_back(free_params_[0]);
-    KinematicsBase::setSearchDiscretization(DEFAULT_SEARCH_DISCRETIZATION);
+    KinematicsBase::setSearchDiscretization(search_discretization);
   }
 
-  urdf::Model robot_model;
-  std::string xml_string;
-
-  std::string urdf_xml, full_urdf_xml;
-  lookupParam("urdf_xml", urdf_xml, robot_description);
-  node_handle.searchParam(urdf_xml, full_urdf_xml);
-
-  ROS_DEBUG_NAMED(name_, "Reading xml file from parameter server");
-  if (!node_handle.getParam(full_urdf_xml, xml_string))
+  const moveit::core::JointModelGroup* jmg = robot_model_->getJointModelGroup(group_name);
+  if (!jmg)
   {
-    ROS_FATAL_NAMED(name_, "Could not load the xml from parameter server: %s", urdf_xml.c_str());
+    ROS_ERROR_STREAM_NAMED(name_, "Unknown planning group: " << group_name);
     return false;
   }
 
-  robot_model.initString(xml_string);
-
-  ROS_DEBUG_STREAM_NAMED(name_, "Reading joints and links from URDF");
-
-  urdf::LinkConstSharedPtr link = robot_model.getLink(getTipFrame());
-  while (link->name != base_frame_ && joint_names_.size() <= num_joints_)
+  ROS_DEBUG_STREAM_NAMED(name_, "Registering joints and links");
+  const moveit::core::LinkModel* link = robot_model_->getLinkModel(tip_frames_[0]);
+  const moveit::core::LinkModel* base_link = robot_model_->getLinkModel(base_frame_);
+  while (link && link != base_link)
   {
-    ROS_DEBUG_NAMED(name_, "Link %s", link->name.c_str());
-    link_names_.push_back(link->name);
-    urdf::JointSharedPtr joint = link->parent_joint;
-    if (joint)
+    ROS_DEBUG_STREAM_NAMED(name_, "Link " << link->getName());
+    link_names_.push_back(link->getName());
+    const moveit::core::JointModel* joint = link->getParentJointModel();
+    if (joint->getType() != joint->UNKNOWN && joint->getType() != joint->FIXED && joint->getVariableCount() == 1)
     {
-      if (joint->type != urdf::Joint::UNKNOWN && joint->type != urdf::Joint::FIXED)
-      {
-        ROS_DEBUG_STREAM_NAMED(name_, "Adding joint " << joint->name);
+      ROS_DEBUG_STREAM_NAMED(name_, "Adding joint " << joint->getName());
 
-        joint_names_.push_back(joint->name);
-        float lower, upper;
-        int hasLimits;
-        if (joint->type != urdf::Joint::CONTINUOUS)
-        {
-          if (joint->safety)
-          {
-            lower = joint->safety->soft_lower_limit;
-            upper = joint->safety->soft_upper_limit;
-          }
-          else
-          {
-            lower = joint->limits->lower;
-            upper = joint->limits->upper;
-          }
-          hasLimits = 1;
-        }
-        else
-        {
-          lower = -M_PI;
-          upper = M_PI;
-          hasLimits = 0;
-        }
-        if (hasLimits)
-        {
-          joint_has_limits_vector_.push_back(true);
-          joint_min_vector_.push_back(lower);
-          joint_max_vector_.push_back(upper);
-        }
-        else
-        {
-          joint_has_limits_vector_.push_back(false);
-          joint_min_vector_.push_back(-M_PI);
-          joint_max_vector_.push_back(M_PI);
-        }
-      }
+      joint_names_.push_back(joint->getName());
+      const moveit::core::VariableBounds& bounds = joint->getVariableBounds()[0];
+      joint_has_limits_vector_.push_back(bounds.position_bounded_);
+      joint_min_vector_.push_back(bounds.min_position_);
+      joint_max_vector_.push_back(bounds.max_position_);
     }
-    else
-    {
-      ROS_WARN_NAMED(name_, "no joint corresponding to %s", link->name.c_str());
-    }
-    link = link->getParent();
+    link = link->getParentLinkModel();
   }
 
   if (joint_names_.size() != num_joints_)
   {
-    ROS_FATAL_STREAM_NAMED(name_, "Joint numbers mismatch: URDF has " << joint_names_.size() << " and IKFast has "
-                                                                      << num_joints_);
+    ROS_FATAL_NAMED(name_, "Joint numbers of RobotModel (%zd) and IKFast solver (%zd) do not match",
+                    joint_names_.size(), num_joints_);
     return false;
   }
 
@@ -449,15 +475,16 @@ bool IKFastKinematicsPlugin::initialize(const std::string& robot_description, co
   std::reverse(joint_max_vector_.begin(), joint_max_vector_.end());
   std::reverse(joint_has_limits_vector_.begin(), joint_has_limits_vector_.end());
 
-  for (size_t i = 0; i < num_joints_; ++i)
-    ROS_DEBUG_STREAM_NAMED(name_, joint_names_[i] << " " << joint_min_vector_[i] << " " << joint_max_vector_[i] << " "
-                                                  << joint_has_limits_vector_[i]);
+  for (size_t joint_id = 0; joint_id < num_joints_; ++joint_id)
+    ROS_DEBUG_STREAM_NAMED(name_, joint_names_[joint_id] << " " << joint_min_vector_[joint_id] << " "
+                                                         << joint_max_vector_[joint_id] << " "
+                                                         << joint_has_limits_vector_[joint_id]);
 
-  active_ = true;
+  initialized_ = true;
   return true;
 }
 
-void IKFastKinematicsPlugin::setSearchDiscretization(const std::map<int, double>& discretization)
+void IKFastKinematicsPlugin::setSearchDiscretization(const std::map<unsigned int, double>& discretization)
 {
   if (discretization.empty())
   {
@@ -467,22 +494,22 @@ void IKFastKinematicsPlugin::setSearchDiscretization(const std::map<int, double>
 
   if (redundant_joint_indices_.empty())
   {
-    ROS_ERROR_STREAM("This group's solver doesn't support redundant joints");
+    ROS_ERROR_STREAM_NAMED(name_, "This group's solver doesn't support redundant joints");
     return;
   }
 
   if (discretization.begin()->first != redundant_joint_indices_[0])
   {
     std::string redundant_joint = joint_names_[free_params_[0]];
-    ROS_ERROR_STREAM("Attempted to discretize a non-redundant joint "
-                     << discretization.begin()->first << ", only joint '" << redundant_joint << "' with index "
-                     << redundant_joint_indices_[0] << " is redundant.");
+    ROS_ERROR_STREAM_NAMED(name_, "Attempted to discretize a non-redundant joint "
+                                      << discretization.begin()->first << ", only joint '" << redundant_joint
+                                      << "' with index " << redundant_joint_indices_[0] << " is redundant.");
     return;
   }
 
   if (discretization.begin()->second <= 0.0)
   {
-    ROS_ERROR_STREAM("Discretization can not takes values that are <= 0");
+    ROS_ERROR_STREAM_NAMED(name_, "Discretization can not takes values that are <= 0");
     return;
   }
 
@@ -492,12 +519,12 @@ void IKFastKinematicsPlugin::setSearchDiscretization(const std::map<int, double>
 
 bool IKFastKinematicsPlugin::setRedundantJoints(const std::vector<unsigned int>& redundant_joint_indices)
 {
-  ROS_ERROR_STREAM("Changing the redundant joints isn't permitted by this group's solver ");
+  ROS_ERROR_STREAM_NAMED(name_, "Changing the redundant joints isn't permitted by this group's solver ");
   return false;
 }
 
-int IKFastKinematicsPlugin::solve(KDL::Frame& pose_frame, const std::vector<double>& vfree,
-                                  IkSolutionList<IkReal>& solutions) const
+size_t IKFastKinematicsPlugin::solve(KDL::Frame& pose_frame, const std::vector<double>& vfree,
+                                     IkSolutionList<IkReal>& solutions) const
 {
   // IKFast56/61
   solutions.Clear();
@@ -530,7 +557,7 @@ int IKFastKinematicsPlugin::solve(KDL::Frame& pose_frame, const std::vector<doub
       vals[8] = mult(2, 2);
 
       // IKFast56/61
-      ComputeIk(trans, vals, vfree.size() > 0 ? &vfree[0] : NULL, solutions);
+      ComputeIk(trans, vals, vfree.size() > 0 ? &vfree[0] : nullptr, solutions);
       return solutions.GetNumSolutions();
 
     case IKP_Direction3D:
@@ -540,7 +567,7 @@ int IKFastKinematicsPlugin::solve(KDL::Frame& pose_frame, const std::vector<doub
       // direction.
 
       direction = pose_frame.M * KDL::Vector(0, 0, 1);
-      ComputeIk(trans, direction.data, vfree.size() > 0 ? &vfree[0] : NULL, solutions);
+      ComputeIk(trans, direction.data, vfree.size() > 0 ? &vfree[0] : nullptr, solutions);
       return solutions.GetNumSolutions();
 
     case IKP_TranslationXAxisAngle4D:
@@ -571,7 +598,7 @@ int IKFastKinematicsPlugin::solve(KDL::Frame& pose_frame, const std::vector<doub
       // direction needs to be orthogonal to z axis and be rotated at a certain angle starting from the x axis (defined
       // in the manipulator base link’s coordinate system)
       pose_frame.M.GetRPY(roll, pitch, yaw);
-      ComputeIk(trans, &yaw, vfree.size() > 0 ? &vfree[0] : NULL, solutions);
+      ComputeIk(trans, &yaw, vfree.size() > 0 ? &vfree[0] : nullptr, solutions);
       return solutions.GetNumSolutions();
 
     case IKP_TranslationYAxisAngleXNorm4D:
@@ -579,7 +606,7 @@ int IKFastKinematicsPlugin::solve(KDL::Frame& pose_frame, const std::vector<doub
       // direction needs to be orthogonal to x axis and be rotated at a certain angle starting from the y axis (defined
       // in the manipulator base link’s coordinate system)
       pose_frame.M.GetRPY(roll, pitch, yaw);
-      ComputeIk(trans, &roll, vfree.size() > 0 ? &vfree[0] : NULL, solutions);
+      ComputeIk(trans, &roll, vfree.size() > 0 ? &vfree[0] : nullptr, solutions);
       return solutions.GetNumSolutions();
 
     case IKP_TranslationZAxisAngleYNorm4D:
@@ -587,7 +614,7 @@ int IKFastKinematicsPlugin::solve(KDL::Frame& pose_frame, const std::vector<doub
       // direction needs to be orthogonal to y axis and be rotated at a certain angle starting from the z axis (defined
       // in the manipulator base link’s coordinate system)
       pose_frame.M.GetRPY(roll, pitch, yaw);
-      ComputeIk(trans, &pitch, vfree.size() > 0 ? &vfree[0] : NULL, solutions);
+      ComputeIk(trans, &pitch, vfree.size() > 0 ? &vfree[0] : nullptr, solutions);
       return solutions.GetNumSolutions();
 
     default:
@@ -606,14 +633,15 @@ void IKFastKinematicsPlugin::getSolution(const IkSolutionList<IkReal>& solutions
   // IKFast56/61
   const IkSolutionBase<IkReal>& sol = solutions.GetSolution(i);
   std::vector<IkReal> vsolfree(sol.GetFree().size());
-  sol.GetSolution(&solution[0], vsolfree.size() > 0 ? &vsolfree[0] : NULL);
+  sol.GetSolution(&solution[0], vsolfree.size() > 0 ? &vsolfree[0] : nullptr);
 
-  // std::cout << "solution " << i << ":" ;
-  // for(int j=0;j<num_joints_; ++j)
-  //   std::cout << " " << solution[j];
-  // std::cout << std::endl;
-
-  // ROS_ERROR("%f %d",solution[2],vsolfree.size());
+  for (std::size_t joint_id = 0; joint_id < num_joints_; ++joint_id)
+  {
+    if (joint_has_limits_vector_[joint_id])
+    {
+      solution[joint_id] = enforceLimits(solution[joint_id], joint_min_vector_[joint_id], joint_max_vector_[joint_id]);
+    }
+  }
 }
 
 void IKFastKinematicsPlugin::getSolution(const IkSolutionList<IkReal>& solutions,
@@ -626,13 +654,14 @@ void IKFastKinematicsPlugin::getSolution(const IkSolutionList<IkReal>& solutions
   // IKFast56/61
   const IkSolutionBase<IkReal>& sol = solutions.GetSolution(i);
   std::vector<IkReal> vsolfree(sol.GetFree().size());
-  sol.GetSolution(&solution[0], vsolfree.size() > 0 ? &vsolfree[0] : NULL);
+  sol.GetSolution(&solution[0], vsolfree.size() > 0 ? &vsolfree[0] : nullptr);
 
   // rotate joints by +/-360° where it is possible and useful
   for (std::size_t i = 0; i < num_joints_; ++i)
   {
     if (joint_has_limits_vector_[i])
     {
+      solution[i] = enforceLimits(solution[i], joint_min_vector_[i], joint_max_vector_[i]);
       double signed_distance = solution[i] - ik_seed_state[i];
       while (signed_distance > M_PI && solution[i] - 2 * M_PI > (joint_min_vector_[i] - LIMIT_TOLERANCE))
       {
@@ -648,82 +677,20 @@ void IKFastKinematicsPlugin::getSolution(const IkSolutionList<IkReal>& solutions
   }
 }
 
-double IKFastKinematicsPlugin::harmonize(const std::vector<double>& ik_seed_state, std::vector<double>& solution) const
+double IKFastKinematicsPlugin::enforceLimits(double joint_value, double min, double max) const
 {
-  double dist_sqr = 0;
-  std::vector<double> ss = ik_seed_state;
-  for (size_t i = 0; i < ik_seed_state.size(); ++i)
+  // If the joint_value is greater than max subtract 2 * PI until it is less than the max
+  while (joint_value > max)
   {
-    while (ss[i] > 2 * M_PI)
-    {
-      ss[i] -= 2 * M_PI;
-    }
-    while (ss[i] < 2 * M_PI)
-    {
-      ss[i] += 2 * M_PI;
-    }
-    while (solution[i] > 2 * M_PI)
-    {
-      solution[i] -= 2 * M_PI;
-    }
-    while (solution[i] < 2 * M_PI)
-    {
-      solution[i] += 2 * M_PI;
-    }
-    dist_sqr += fabs(ik_seed_state[i] - solution[i]);
+    joint_value -= 2 * M_PI;
   }
-  return dist_sqr;
-}
 
-// void IKFastKinematicsPlugin::getOrderedSolutions(const std::vector<double> &ik_seed_state,
-//                                  std::vector<std::vector<double> >& solslist)
-// {
-//   std::vector<double>
-//   double mindist = 0;
-//   int minindex = -1;
-//   std::vector<double> sol;
-//   for(size_t i=0;i<solslist.size();++i){
-//     getSolution(i,sol);
-//     double dist = harmonize(ik_seed_state, sol);
-//     //std::cout << "dist[" << i << "]= " << dist << std::endl;
-//     if(minindex == -1 || dist<mindist){
-//       minindex = i;
-//       mindist = dist;
-//     }
-//   }
-//   if(minindex >= 0){
-//     getSolution(minindex,solution);
-//     harmonize(ik_seed_state, solution);
-//     index = minindex;
-//   }
-// }
-
-void IKFastKinematicsPlugin::getClosestSolution(const IkSolutionList<IkReal>& solutions,
-                                                const std::vector<double>& ik_seed_state,
-                                                std::vector<double>& solution) const
-{
-  double mindist = DBL_MAX;
-  int minindex = -1;
-  std::vector<double> sol;
-
-  // IKFast56/61
-  for (size_t i = 0; i < solutions.GetNumSolutions(); ++i)
+  // If the joint_value is less than the min, add 2 * PI until it is more than the min
+  while (joint_value < min)
   {
-    getSolution(solutions, i, sol);
-    double dist = harmonize(ik_seed_state, sol);
-    ROS_INFO_STREAM_NAMED(name_, "Dist " << i << " dist " << dist);
-    // std::cout << "dist[" << i << "]= " << dist << std::endl;
-    if (minindex == -1 || dist < mindist)
-    {
-      minindex = i;
-      mindist = dist;
-    }
+    joint_value += 2 * M_PI;
   }
-  if (minindex >= 0)
-  {
-    getSolution(solutions, minindex, solution);
-    harmonize(ik_seed_state, solution);
-  }
+  return joint_value;
 }
 
 void IKFastKinematicsPlugin::fillFreeParams(int count, int* array)
@@ -820,7 +787,7 @@ bool IKFastKinematicsPlugin::getPositionFK(const std::vector<std::string>& link_
     p_out.M.data[i] = eerot[i];
 
   poses.resize(1);
-  tf::poseKDLToMsg(p_out, poses[0]);
+  poses[0] = tf2::toMsg(p_out);
 
   return valid;
 }
@@ -830,10 +797,8 @@ bool IKFastKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose& ik_pose
                                               std::vector<double>& solution, moveit_msgs::MoveItErrorCodes& error_code,
                                               const kinematics::KinematicsQueryOptions& options) const
 {
-  const IKCallbackFn solution_callback = 0;
   std::vector<double> consistency_limits;
-
-  return searchPositionIK(ik_pose, ik_seed_state, timeout, consistency_limits, solution, solution_callback, error_code,
+  return searchPositionIK(ik_pose, ik_seed_state, timeout, consistency_limits, solution, IKCallbackFn(), error_code,
                           options);
 }
 
@@ -843,8 +808,7 @@ bool IKFastKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose& ik_pose
                                               std::vector<double>& solution, moveit_msgs::MoveItErrorCodes& error_code,
                                               const kinematics::KinematicsQueryOptions& options) const
 {
-  const IKCallbackFn solution_callback = 0;
-  return searchPositionIK(ik_pose, ik_seed_state, timeout, consistency_limits, solution, solution_callback, error_code,
+  return searchPositionIK(ik_pose, ik_seed_state, timeout, consistency_limits, solution, IKCallbackFn(), error_code,
                           options);
 }
 
@@ -866,9 +830,7 @@ bool IKFastKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose& ik_pose
                                               moveit_msgs::MoveItErrorCodes& error_code,
                                               const kinematics::KinematicsQueryOptions& options) const
 {
-  ROS_DEBUG_STREAM_NAMED(name_, "searchPositionIK");
-
-  /// search_mode is currently fixed during code generation
+  // "SEARCH_MODE" is fixed during code generation
   SEARCH_MODE search_mode = OPTIMIZE_MAX_JOINT;
 
   // Check if there are no redundant joints
@@ -879,7 +841,7 @@ bool IKFastKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose& ik_pose
     std::vector<geometry_msgs::Pose> ik_poses(1, ik_pose);
     std::vector<std::vector<double>> solutions;
     kinematics::KinematicsResult kinematic_result;
-    // Find all IK solution within joint limits
+    // Find all IK solutions within joint limits
     if (!getPositionIK(ik_poses, ik_seed_state, solutions, kinematic_result, options))
     {
       ROS_DEBUG_STREAM_NAMED(name_, "No solution whatsoever");
@@ -928,7 +890,7 @@ bool IKFastKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose& ik_pose
 
   // -------------------------------------------------------------------------------------------------
   // Error Checking
-  if (!active_)
+  if (!initialized_)
   {
     ROS_ERROR_STREAM_NAMED(name_, "Kinematics not active");
     error_code.val = error_code.NO_IK_SOLUTION;
@@ -955,11 +917,10 @@ bool IKFastKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose& ik_pose
   // Initialize
 
   KDL::Frame frame;
-  tf::poseMsgToKDL(ik_pose, frame);
+  transformToChainFrame(ik_pose, frame);
 
   std::vector<double> vfree(free_params_.size());
 
-  ros::Time maxTime = ros::Time::now() + ros::Duration(timeout);
   int counter = 0;
 
   double initial_guess = ik_seed_state[free_params_[0]];
@@ -969,21 +930,22 @@ bool IKFastKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose& ik_pose
   // Handle consitency limits if needed
   int num_positive_increments;
   int num_negative_increments;
+  double search_discretization = redundant_joint_discretization_.at(free_params_[0]);
 
   if (!consistency_limits.empty())
   {
-    // moveit replaced consistency_limit (scalar) w/ consistency_limits (vector)
+    // MoveIt! replaced consistency_limit (scalar) w/ consistency_limits (vector)
     // Assume [0]th free_params element for now.  Probably wrong.
     double max_limit = fmin(joint_max_vector_[free_params_[0]], initial_guess + consistency_limits[free_params_[0]]);
     double min_limit = fmax(joint_min_vector_[free_params_[0]], initial_guess - consistency_limits[free_params_[0]]);
 
-    num_positive_increments = (int)((max_limit - initial_guess) / search_discretization_);
-    num_negative_increments = (int)((initial_guess - min_limit) / search_discretization_);
+    num_positive_increments = static_cast<int>((max_limit - initial_guess) / search_discretization);
+    num_negative_increments = static_cast<int>((initial_guess - min_limit) / search_discretization);
   }
-  else  // no consitency limits provided
+  else  // no consistency limits provided
   {
-    num_positive_increments = (joint_max_vector_[free_params_[0]] - initial_guess) / search_discretization_;
-    num_negative_increments = (initial_guess - joint_min_vector_[free_params_[0]]) / search_discretization_;
+    num_positive_increments = (joint_max_vector_[free_params_[0]] - initial_guess) / search_discretization;
+    num_negative_increments = (initial_guess - joint_min_vector_[free_params_[0]]) / search_discretization;
   }
 
   // -------------------------------------------------------------------------------------------------
@@ -1002,22 +964,20 @@ bool IKFastKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose& ik_pose
   while (true)
   {
     IkSolutionList<IkReal> solutions;
-    int numsol = solve(frame, vfree, solutions);
+    size_t numsol = solve(frame, vfree, solutions);
 
     ROS_DEBUG_STREAM_NAMED(name_, "Found " << numsol << " solutions from IKFast");
 
-    // ROS_INFO("%f",vfree[0]);
-
     if (numsol > 0)
     {
-      for (int s = 0; s < numsol; ++s)
+      for (size_t s = 0; s < numsol; ++s)
       {
         nattempts++;
         std::vector<double> sol;
         getSolution(solutions, ik_seed_state, s, sol);
 
         bool obeys_limits = true;
-        for (unsigned int i = 0; i < sol.size(); i++)
+        for (size_t i = 0; i < sol.size(); i++)
         {
           if (joint_has_limits_vector_[i] && (sol[i] < joint_min_vector_[i] || sol[i] > joint_max_vector_[i]))
           {
@@ -1075,7 +1035,7 @@ bool IKFastKinematicsPlugin::searchPositionIK(const geometry_msgs::Pose& ik_pose
       break;
     }
 
-    vfree[0] = initial_guess + search_discretization_ * counter;
+    vfree[0] = initial_guess + search_discretization * counter;
     // ROS_DEBUG_STREAM_NAMED(name_,"Attempt " << counter << " with 0th free joint having value " << vfree[0]);
   }
 
@@ -1100,16 +1060,16 @@ bool IKFastKinematicsPlugin::getPositionIK(const geometry_msgs::Pose& ik_pose, c
 {
   ROS_DEBUG_STREAM_NAMED(name_, "getPositionIK");
 
-  if (!active_)
+  if (!initialized_)
   {
-    ROS_ERROR("kinematics not active");
+    ROS_ERROR_NAMED(name_, "kinematics not active");
     return false;
   }
 
   if (ik_seed_state.size() < num_joints_)
   {
-    ROS_ERROR_STREAM("ik_seed_state only has " << ik_seed_state.size() << " entries, this ikfast solver requires "
-                                               << num_joints_);
+    ROS_ERROR_STREAM_NAMED(name_, "ik_seed_state only has " << ik_seed_state.size()
+                                                            << " entries, this ikfast solver requires " << num_joints_);
     return false;
   }
 
@@ -1120,9 +1080,10 @@ bool IKFastKinematicsPlugin::getPositionIK(const geometry_msgs::Pose& ik_pose, c
     if (joint_has_limits_vector_[i] && ((ik_seed_state[i] < (joint_min_vector_[i] - LIMIT_TOLERANCE)) ||
                                         (ik_seed_state[i] > (joint_max_vector_[i] + LIMIT_TOLERANCE))))
     {
-      ROS_DEBUG_STREAM_NAMED("ikseed", "Not in limits! " << (int)i << " value " << ik_seed_state[i]
-                                                         << " has limit: " << joint_has_limits_vector_[i] << "  being  "
-                                                         << joint_min_vector_[i] << " to " << joint_max_vector_[i]);
+      ROS_DEBUG_STREAM_NAMED(name_, "IK seed not in limits! " << static_cast<int>(i) << " value " << ik_seed_state[i]
+                                                              << " has limit: " << joint_has_limits_vector_[i]
+                                                              << "  being  " << joint_min_vector_[i] << " to "
+                                                              << joint_max_vector_[i]);
       return false;
     }
   }
@@ -1136,10 +1097,10 @@ bool IKFastKinematicsPlugin::getPositionIK(const geometry_msgs::Pose& ik_pose, c
   }
 
   KDL::Frame frame;
-  tf::poseMsgToKDL(ik_pose, frame);
+  transformToChainFrame(ik_pose, frame);
 
   IkSolutionList<IkReal> solutions;
-  int numsol = solve(frame, vfree, solutions);
+  size_t numsol = solve(frame, vfree, solutions);
   ROS_DEBUG_STREAM_NAMED(name_, "Found " << numsol << " solutions from IKFast");
 
   std::vector<LimitObeyingSol> solutions_obey_limits;
@@ -1151,8 +1112,8 @@ bool IKFastKinematicsPlugin::getPositionIK(const geometry_msgs::Pose& ik_pose, c
     {
       std::vector<double> sol;
       getSolution(solutions, ik_seed_state, s, sol);
-      ROS_DEBUG_NAMED(name_, "Sol %d: %e   %e   %e   %e   %e   %e", (int)s, sol[0], sol[1], sol[2], sol[3], sol[4],
-                      sol[5]);
+      ROS_DEBUG_NAMED(name_, "Sol %d: %e   %e   %e   %e   %e   %e", static_cast<int>(s), sol[0], sol[1], sol[2], sol[3],
+                      sol[4], sol[5]);
 
       bool obeys_limits = true;
       for (std::size_t i = 0; i < sol.size(); i++)
@@ -1163,9 +1124,10 @@ bool IKFastKinematicsPlugin::getPositionIK(const geometry_msgs::Pose& ik_pose, c
         {
           // One element of solution is not within limits
           obeys_limits = false;
-          ROS_DEBUG_STREAM_NAMED(name_, "Not in limits! " << (int)i << " value " << sol[i] << " has limit: "
-                                                          << joint_has_limits_vector_[i] << "  being  "
-                                                          << joint_min_vector_[i] << " to " << joint_max_vector_[i]);
+          ROS_DEBUG_STREAM_NAMED(name_, "Not in limits! " << static_cast<int>(i) << " value " << sol[i]
+                                                          << " has limit: " << joint_has_limits_vector_[i]
+                                                          << "  being  " << joint_min_vector_[i] << " to "
+                                                          << joint_max_vector_[i]);
           break;
         }
       }
@@ -1209,36 +1171,36 @@ bool IKFastKinematicsPlugin::getPositionIK(const std::vector<geometry_msgs::Pose
 {
   ROS_DEBUG_STREAM_NAMED(name_, "getPositionIK with multiple solutions");
 
-  if (!active_)
+  if (!initialized_)
   {
-    ROS_ERROR("kinematics not active");
+    ROS_ERROR_NAMED(name_, "kinematics not active");
     result.kinematic_error = kinematics::KinematicErrors::SOLVER_NOT_ACTIVE;
     return false;
   }
 
   if (ik_poses.empty())
   {
-    ROS_ERROR("ik_poses is empty");
+    ROS_ERROR_NAMED(name_, "ik_poses is empty");
     result.kinematic_error = kinematics::KinematicErrors::EMPTY_TIP_POSES;
     return false;
   }
 
   if (ik_poses.size() > 1)
   {
-    ROS_ERROR("ik_poses contains multiple entries, only one is allowed");
+    ROS_ERROR_NAMED(name_, "ik_poses contains multiple entries, only one is allowed");
     result.kinematic_error = kinematics::KinematicErrors::MULTIPLE_TIPS_NOT_SUPPORTED;
     return false;
   }
 
   if (ik_seed_state.size() < num_joints_)
   {
-    ROS_ERROR_STREAM("ik_seed_state only has " << ik_seed_state.size() << " entries, this ikfast solver requires "
-                                               << num_joints_);
+    ROS_ERROR_STREAM_NAMED(name_, "ik_seed_state only has " << ik_seed_state.size()
+                                                            << " entries, this ikfast solver requires " << num_joints_);
     return false;
   }
 
   KDL::Frame frame;
-  tf::poseMsgToKDL(ik_poses[0], frame);
+  transformToChainFrame(ik_poses[0], frame);
 
   // solving ik
   std::vector<IkSolutionList<IkReal>> solution_set;
@@ -1262,7 +1224,7 @@ bool IKFastKinematicsPlugin::getPositionIK(const std::vector<geometry_msgs::Pose
       if (!((jv > (joint_min - LIMIT_TOLERANCE)) && (jv < (joint_max + LIMIT_TOLERANCE))))
       {
         result.kinematic_error = kinematics::KinematicErrors::IK_SEED_OUTSIDE_LIMITS;
-        ROS_ERROR_STREAM("ik seed is out of bounds");
+        ROS_ERROR_STREAM_NAMED(name_, "ik seed is out of bounds");
         return false;
       }
     }
@@ -1347,23 +1309,17 @@ bool IKFastKinematicsPlugin::getPositionIK(const std::vector<geometry_msgs::Pose
 bool IKFastKinematicsPlugin::sampleRedundantJoint(kinematics::DiscretizationMethod method,
                                                   std::vector<double>& sampled_joint_vals) const
 {
-  double joint_min = -M_PI;
-  double joint_max = M_PI;
   int index = redundant_joint_indices_.front();
   double joint_dscrt = redundant_joint_discretization_.at(index);
-
-  if (joint_has_limits_vector_[redundant_joint_indices_.front()])
-  {
-    joint_min = joint_min_vector_[index];
-    joint_max = joint_max_vector_[index];
-  }
+  double joint_min = joint_min_vector_[index];
+  double joint_max = joint_max_vector_[index];
 
   switch (method)
   {
     case kinematics::DiscretizationMethods::ALL_DISCRETIZED:
     {
-      int steps = std::ceil((joint_max - joint_min) / joint_dscrt);
-      for (unsigned int i = 0; i < steps; i++)
+      size_t steps = std::ceil((joint_max - joint_min) / joint_dscrt);
+      for (size_t i = 0; i < steps; i++)
       {
         sampled_joint_vals.push_back(joint_min + joint_dscrt * i);
       }
@@ -1386,15 +1342,35 @@ bool IKFastKinematicsPlugin::sampleRedundantJoint(kinematics::DiscretizationMeth
 
       break;
     default:
-      ROS_ERROR_STREAM("Discretization method " << method << " is not supported");
+      ROS_ERROR_STREAM_NAMED(name_, "Discretization method " << method << " is not supported");
       return false;
   }
 
   return true;
 }
 
+void IKFastKinematicsPlugin::transformToChainFrame(const geometry_msgs::Pose& ik_pose, KDL::Frame& ik_pose_chain) const
+{
+  if (tip_transform_required_ || base_transform_required_)
+  {
+    Eigen::Isometry3d ik_eigen_pose;
+    tf2::fromMsg(ik_pose, ik_eigen_pose);
+    if (tip_transform_required_)
+      ik_eigen_pose = ik_eigen_pose * group_tip_to_chain_tip_;
+
+    if (base_transform_required_)
+      ik_eigen_pose = chain_base_to_group_base_ * ik_eigen_pose;
+
+    tf::transformEigenToKDL(ik_eigen_pose, ik_pose_chain);
+  }
+  else
+  {
+    tf2::fromMsg(ik_pose, ik_pose_chain);
+  }
+}
+
 }  // end namespace
 
 // register IKFastKinematicsPlugin as a KinematicsBase implementation
 #include <pluginlib/class_list_macros.hpp>
-PLUGINLIB_EXPORT_CLASS(ikfast_kinematics_plugin::IKFastKinematicsPlugin, kinematics::KinematicsBase);
+PLUGINLIB_EXPORT_CLASS(iiwa7_iiwa_arm::IKFastKinematicsPlugin, kinematics::KinematicsBase);
